fix(offline): share one ios navigation provider (fixes two-cores crash)
iOS Nav SDK allows only ONE active MapboxNavigationProvider — the offline module
allocated its own while the embedded view had one, tripping the SDK's
"Two simultaneous active navigation cores" [BUG] abort the moment the en-route
screen fired the corridor pre-download (crash on entering en-route).

Add SharedNavigationProvider (a single app-wide provider, created on first use)
and route BOTH the view's embed() and the offline module through it — mirroring
Android's MapboxNavigationProvider process singleton (which is why Android never
hit this). The view still holds/niles its ref per session; the singleton retains
the core. Offline-reroute config (.hybrid + tilestoreConfig .default) lives in
the singleton.

Native-only (Swift) — no lib change. Bump 0.6.0 -> 0.6.1.
